+++ R/do.R
@@ -1,8 +1,16 @@
-#' A convenience function that either: a) applies a function to an input or b) runs a parameterless function. A useful shorthand alternative to piping into an anonymous function (a) or manually removing intermediate objects used in creating an output (b).
-#' @param x An object passed into `f`. If left NULL (default) then `f` will be called without parameters.
-#' @param f A function to apply on `x` or ran without parameters (if `x` is NULL).
+#' A convenience function that applies a function to an input, a more readable shorthand for "checkpointing" with an anonymous function.
+#' @param x An object passed into `f`.
+#' @param f A function to apply on `x`.
 #' @return The output of `f(x)`, whatever that may be.
 #' @export
-do = function(x=NULL, f){
-  if (is.null(x)) f() else f(x)
+do = function(x, f){
+  return(f(x))
+}
+
+#' A convenience function performs a no-argument function, a useful shorthand for using a functional environment to prevent intermediate object leakage into the main environment.
+#' @param f A no-argument function to perform.
+#' @return The output of `f()`, whatever that may be.
+#' @export
+do0 = function(f){
+  return(f())
 }