+++ R/setXNames.R
@@ -4,8 +4,8 @@
 #' @return A copy of `object` with its row names changed to `nm`.
 #' @export
 setRownames = function(object, nm){
-  rownames(object)=nm
-  object
+  rownames(object) = nm
+  return(object)
 }
 
 #' A convenience function to modify the column names of the passed in object.
@@ -14,16 +14,16 @@ setRownames = function(object, nm){
 #' @return A copy of `object` with its column names changed to `nm`.
 #' @export
 setColnames = function(object, nm){
-  colnames(object)=nm
-  object
+  colnames(object) = nm
+  return(object)
 }
 
 #' A convenience function to modify the dimension names of the passed in object.
 #' @param object The object whose dimension names will be modified.
-#' @param nm A list of character vectors containing the new dimension names for `object`.
-#' @return A copy of `object` with its dimension changed to the elements of `nm`.
+#' @param ... Character vectors containing the new dimension names for `object`.
+#' @return A copy of `object` with its dimension changed to the elements of `...`.
 #' @export
-setDimnames = function(object, nm){
-  dimnames(object)=nm
-  object
+setDimnames = function(object, ...){
+  dimnames(object) = list(...)
+  return(object)
 }